refactor to extract load goosehints to a separate file
