+++ crates/goose-mcp/src/developer/mod.rs
@@ -1,4 +1,5 @@
 mod editor_models;
+mod goose_hints;
 mod lang;
 mod shell;
 
@@ -409,53 +410,7 @@ impl DeveloperRouter {
             .and_then(|s| serde_json::from_str(&s).ok())
             .unwrap_or_else(|| vec![".goosehints".to_string()]);
 
-        let mut global_hints_contents = Vec::with_capacity(hints_filenames.len());
-        let mut local_hints_contents = Vec::with_capacity(hints_filenames.len());
-
-        for hints_filename in &hints_filenames {
-            // Global hints
-            // choose_app_strategy().config_dir()
-            // - macOS/Linux: ~/.config/goose/
-            // - Windows:     ~\AppData\Roaming\Block\goose\config\
-            // keep previous behavior of expanding ~/.config in case this fails
-            let global_hints_path = choose_app_strategy(crate::APP_STRATEGY.clone())
-                .map(|strategy| strategy.in_config_dir(hints_filename))
-                .unwrap_or_else(|_| {
-                    let path_str = format!("~/.config/goose/{}", hints_filename);
-                    PathBuf::from(shellexpand::tilde(&path_str).to_string())
-                });
-
-            if let Some(parent) = global_hints_path.parent() {
-                let _ = std::fs::create_dir_all(parent);
-            }
-
-            if global_hints_path.is_file() {
-                if let Ok(content) = std::fs::read_to_string(&global_hints_path) {
-                    global_hints_contents.push(content);
-                }
-            }
-
-            let local_hints_path = cwd.join(hints_filename);
-            if local_hints_path.is_file() {
-                if let Ok(content) = std::fs::read_to_string(&local_hints_path) {
-                    local_hints_contents.push(content);
-                }
-            }
-        }
-
-        let mut hints = String::new();
-        if !global_hints_contents.is_empty() {
-            hints.push_str("\n### Global Hints\nThe developer extension includes some global hints that apply to all projects & directories.\n");
-            hints.push_str(&global_hints_contents.join("\n"));
-        }
-
-        if !local_hints_contents.is_empty() {
-            if !hints.is_empty() {
-                hints.push_str("\n\n");
-            }
-            hints.push_str("### Project Hints\nThe developer extension includes some hints for working on the project in this directory.\n");
-            hints.push_str(&local_hints_contents.join("\n"));
-        }
+        let hints = goose_hints::load_and_format_hints(&cwd, &hints_filenames);
 
         // Return base instructions directly when no hints are found
         let instructions = if hints.is_empty() {
@@ -1577,68 +1532,12 @@ mod tests {
     use core::panic;
     use serde_json::json;
     use serial_test::serial;
-    use std::fs::{self, read_to_string};
+    use std::fs::read_to_string;
     use tempfile::TempDir;
     use tokio::sync::OnceCell;
 
-    #[test]
-    #[serial]
-    fn test_global_goosehints() {
-        // if ~/.config/goose/.goosehints exists, it should be included in the instructions
-        // copy the existing global hints file to a .bak file
-        let global_hints_path =
-            PathBuf::from(shellexpand::tilde("~/.config/goose/.goosehints").to_string());
-        let global_hints_bak_path =
-            PathBuf::from(shellexpand::tilde("~/.config/goose/.goosehints.bak").to_string());
-        let mut globalhints_existed = false;
 
-        if global_hints_path.is_file() {
-            globalhints_existed = true;
-            fs::copy(&global_hints_path, &global_hints_bak_path).unwrap();
-        }
 
-        fs::write(&global_hints_path, "These are my global goose hints.").unwrap();
-
-        let dir = TempDir::new().unwrap();
-        std::env::set_current_dir(dir.path()).unwrap();
-
-        let router = DeveloperRouter::new();
-        let instructions = router.instructions();
-
-        assert!(instructions.contains("### Global Hints"));
-        assert!(instructions.contains("my global goose hints."));
-
-        // restore backup if globalhints previously existed
-        if globalhints_existed {
-            fs::copy(&global_hints_bak_path, &global_hints_path).unwrap();
-            fs::remove_file(&global_hints_bak_path).unwrap();
-        }
-    }
-
-    #[test]
-    #[serial]
-    fn test_goosehints_when_present() {
-        let dir = TempDir::new().unwrap();
-        std::env::set_current_dir(dir.path()).unwrap();
-
-        fs::write(".goosehints", "Test hint content").unwrap();
-        let router = DeveloperRouter::new();
-        let instructions = router.instructions();
-
-        assert!(instructions.contains("Test hint content"));
-    }
-
-    #[test]
-    #[serial]
-    fn test_goosehints_when_missing() {
-        let dir = TempDir::new().unwrap();
-        std::env::set_current_dir(dir.path()).unwrap();
-
-        let router = DeveloperRouter::new();
-        let instructions = router.instructions();
-
-        assert!(!instructions.contains("Project Hints"));
-    }
 
     static DEV_ROUTER: OnceCell<DeveloperRouter> = OnceCell::const_new();
 
@@ -1668,38 +1567,7 @@ mod tests {
         temp_dir.close().unwrap();
     }
 
-    #[test]
-    #[serial]
-    fn test_goosehints_multiple_filenames() {
-        let dir = TempDir::new().unwrap();
-        std::env::set_current_dir(dir.path()).unwrap();
-        std::env::set_var("CONTEXT_FILE_NAMES", r#"["CLAUDE.md", ".goosehints"]"#);
 
-        fs::write("CLAUDE.md", "Custom hints file content from CLAUDE.md").unwrap();
-        fs::write(".goosehints", "Custom hints file content from .goosehints").unwrap();
-        let router = DeveloperRouter::new();
-        let instructions = router.instructions();
-
-        assert!(instructions.contains("Custom hints file content from CLAUDE.md"));
-        assert!(instructions.contains("Custom hints file content from .goosehints"));
-        std::env::remove_var("CONTEXT_FILE_NAMES");
-    }
-
-    #[test]
-    #[serial]
-    fn test_goosehints_configurable_filename() {
-        let dir = TempDir::new().unwrap();
-        std::env::set_current_dir(dir.path()).unwrap();
-        std::env::set_var("CONTEXT_FILE_NAMES", r#"["CLAUDE.md"]"#);
-
-        fs::write("CLAUDE.md", "Custom hints file content").unwrap();
-        let router = DeveloperRouter::new();
-        let instructions = router.instructions();
-
-        assert!(instructions.contains("Custom hints file content"));
-        assert!(!instructions.contains(".goosehints")); // Make sure it's not loading the default
-        std::env::remove_var("CONTEXT_FILE_NAMES");
-    }
 
     #[tokio::test]
     #[serial]
